+++ packages/react-server/src/server/server.ts
@@ -33,12 +33,18 @@ interface Options {
 export function createServer(options: Options): Server {
   const {
     /* eslint-disable no-process-env */
-    ip = process.env.REACT_SERVER_IP || 'localhost',
-    port = (process.env.REACT_SERVER_PORT &&
-      parseInt(process.env.REACT_SERVER_PORT, 10)) ||
-      8081,
+    ip = process.env.REACT_SERVER_IP &&
+    process.env.REACT_SERVER_IP !== 'undefined'
+      ? process.env.REACT_SERVER_IP
+      : '0.0.0.0',
+    port = process.env.REACT_SERVER_PORT &&
+    process.env.REACT_SERVER_PORT !== 'undefined'
+      ? parseInt(process.env.REACT_SERVER_PORT, 10)
+      : 8081,
     // a default is set in sewingKitMiddleware
-    assetPrefix = process.env.CDN_URL,
+    assetPrefix = process.env.CDN_URL && process.env.CDN_URL !== 'undefined'
+      ? process.env.CDN_URL
+      : undefined,
     /* eslint-enable no-process-env */
     render,
     renderError,
@@ -61,7 +67,7 @@ export function createServer(options: Options): Server {
 
   app.use(createRender(render, {assetPrefix, assetName, renderError}));
 
-  return app.listen(port || 3000, ip, () => {
+  return app.listen(port, ip, () => {
     logger.log(`started react-server on ${ip}:${port}`);
   });
 }